+++ pagegraph/graph/__init__.py
@@ -5,31 +5,37 @@ from typing import Any, cast, Dict, Iterable, List, Tuple
 import networkx as NWX  # type: ignore
 
 from pagegraph.graph.edge import Edge, NodeInsertEdge, JSCallEdge
+from pagegraph.graph.edge import RequestStartEdge
 from pagegraph.graph.edge import for_type as edge_for_type
 from pagegraph.graph.node import for_type as node_for_type
 from pagegraph.graph.node import DOMRootNode, Node, HTMLNode, ScriptNode
 from pagegraph.graph.node import ParserNode, FrameOwnerNode, ResourceNode
 from pagegraph.graph.node import TextNode, JSStructureNode
-from pagegraph.graph.types import BlinkId, NodeIterator, PageGraphId, DOMNode
-from pagegraph.graph.types import ChildNode, ParentNode, EdgeIterator, FrameId
+from pagegraph.graph.requests import RequestChain, request_chain_for_edge
+from pagegraph.types import BlinkId, NodeIterator, PageGraphId, DOMNode
+from pagegraph.types import ChildNode, ParentNode, EdgeIterator, FrameId
+from pagegraph.types import RequestId
 from pagegraph.util import check_pagegraph_version
 
 
 class PageGraph:
 
+    # Instance properties
     debug: bool = True
     graph: NWX.MultiDiGraph
     r_graph: NWX.MultiDiGraph
-    blink_id_mapping: Dict[BlinkId, DOMNode] = {}
 
-    nodes_by_type: Dict[Node.Types, List[Node]] = {}
-    edges_by_type: Dict[Edge.Types, List[Edge]] = {}
-    edge_cache: List[Edge] = []
-    edge_id_cache: Dict[PageGraphId, Tuple[Any, Any]] = {}
+    __blink_id_map: Dict[BlinkId, DOMNode] = {}
+    __request_chain_map: Dict[RequestId, RequestChain] = {}
 
-    inserted_below_mapping: Dict[ParentNode, List[ChildNode]] = {}
+    __nodes_by_type: Dict[Node.Types, List[Node]] = {}
+    __edges_by_type: Dict[Edge.Types, List[Edge]] = {}
+    __edge_cache: List[Edge] = []
+    __edge_id_cache: Dict[PageGraphId, Tuple[Any, Any]] = {}
+
+    __inserted_below_map: Dict[ParentNode, List[ChildNode]] = {}
     # Mapping from a frame id to the most recent DOM node seen for the frame
-    frame_id_mapping: Dict[FrameId, DOMRootNode] = {}
+    __frame_id_map: Dict[FrameId, DOMRootNode] = {}
 
     def __init__(self, graph: NWX.MultiDiGraph, debug: bool = False):
         self.graph = graph
@@ -40,61 +46,78 @@ class PageGraph:
         # and the frame_id to frame node mapping (we keep the most
         # recent version of each frame).
         for node_type in Node.Types:
-            self.nodes_by_type[node_type] = []
+            self.__nodes_by_type[node_type] = []
         for edge_type in Edge.Types:
-            self.edges_by_type[edge_type] = []
+            self.__edges_by_type[edge_type] = []
 
         for node in self.nodes():
-            self.nodes_by_type[node.node_type()].append(node)
+            self.__nodes_by_type[node.node_type()].append(node)
         for edge in self.edges():
-            self.edges_by_type[edge.edge_type()].append(edge)
+            self.__edges_by_type[edge.edge_type()].append(edge)
 
-        # We iterate over nodes a second time if we're in debug mode
-        # because some internal checks are dependent on caches
-        # having already seen other nodes, and so this removes those cycles.
-        if self.debug:
-            for node in self.nodes():
+        # Now that we've mapped the NetworkX representation of the
+        # graph into the representation defined by this library,
+        # we iterate over each node and edge again, to allow
+        # each element to take steps and operations only
+        # safe to do once we know the entire graph is hooked up.
+        for node in self.nodes():
+            if self.debug:
                 node.validate()
+            node.build_caches()
+        for edge in self.edges():
+            if self.debug:
+                edge.validate()
+            edge.build_caches()
 
         for node in self.dom_nodes():
             if node.is_domroot():
                 domroot_node = cast(DOMRootNode, node)
                 blink_id = domroot_node.blink_id()
-                if blink_id not in self.frame_id_mapping:
-                    self.frame_id_mapping[blink_id] = domroot_node
+                if blink_id not in self.__frame_id_map:
+                    self.__frame_id_map[blink_id] = domroot_node
                 else:
-                    current_node = self.frame_id_mapping[blink_id]
+                    current_node = self.__frame_id_map[blink_id]
                     if current_node.timestamp() < domroot_node.timestamp():
-                        self.frame_id_mapping[blink_id] = domroot_node
+                        self.__frame_id_map[blink_id] = domroot_node
 
-        # Finally, populate the mapping of what nodes were ever inserted
-        # below another node, during the document's lifetime
-        for edge in self.insert_edges():
-            pass
+        for request_start_edge in self.request_start_edges():
+            request_id = request_start_edge.request_id()
+            self.__request_chain_map[request_id] = request_chain_for_edge(
+                request_start_edge)
+
+    def request_chain_for_id(self, request_id: RequestId) -> RequestChain:
+        if self.debug:
+            if request_id not in self.__request_chain_map:
+                raise Exception(f"Unrecognized request id: {request_id}")
+        return self.__request_chain_map[request_id]
 
     def nodes(self) -> list[Node]:
         return [self.node(node_id) for node_id in self.graph.nodes()]
 
     def edges(self) -> EdgeIterator:
-        if len(self.edge_cache) > 0:
-            return self.edge_cache
+        if len(self.__edge_cache) > 0:
+            return self.__edge_cache
         edges = []
         for u, v, edge_id in self.graph.edges:
-            self.edge_id_cache[edge_id] = (u, v)
+            self.__edge_id_cache[edge_id] = (u, v)
             edges.append(self.edge(edge_id))
-        self.edge_cache = edges
+        self.__edge_cache = edges
         return edges
 
-    def insert_edges(self) -> Iterable[NodeInsertEdge]:
-        return [cast(NodeInsertEdge, e) for e in
-                self.edges_of_type(Edge.Types.NODE_INSERT)]
+    def insert_edges(self) -> List[NodeInsertEdge]:
+        edges = self.edges_of_type(Edge.Types.NODE_INSERT)
+        return cast(List[NodeInsertEdge], edges)
+
+    def request_start_edges(self) -> List[RequestStartEdge]:
+        edges = self.edges_of_type(Edge.Types.REQUEST_START)
+        return cast(List[RequestStartEdge], edges)
 
     def node_for_blink_id(self, blink_id: BlinkId) -> Node:
         if self.debug:
-            if blink_id not in self.blink_id_mapping:
+            if blink_id not in self.__blink_id_map:
                 raise Exception(f"blink_id not in blink_id cache: {blink_id}")
 
-        node = self.blink_id_mapping[blink_id]
+        node = self.__blink_id_map[blink_id]
         if node.is_domroot():
             return cast(DOMRootNode, node)
 
@@ -115,16 +138,16 @@ class PageGraph:
         return cast(List[DOMNode], nodes)
 
     def nodes_of_type(self, node_type: Node.Types) -> List[Node]:
-        return self.nodes_by_type[node_type]
+        return self.__nodes_by_type[node_type]
 
     def edges_of_type(self, edge_type: Edge.Types) -> List[Edge]:
-        return self.edges_by_type[edge_type]
+        return self.__edges_by_type[edge_type]
 
     def domroot_for_frame_id(self, frame_id: FrameId) -> DOMRootNode:
         if self.debug:
-            if frame_id not in self.frame_id_mapping:
-                raise Exception(f"frame_id not in frame_id_mapping:{frame_id}")
-        return self.frame_id_mapping[frame_id]
+            if frame_id not in self.__frame_id_map:
+                raise Exception(f"frame_id not in __frame_id_map:{frame_id}")
+        return self.__frame_id_map[frame_id]
 
     def resource_nodes(self) -> Iterable[ResourceNode]:
         node_iterator = self.nodes_of_type(Node.Types.RESOURCE)
@@ -164,9 +187,9 @@ class PageGraph:
                         parent_node: ParentNode) -> List[ChildNode] | None:
         """Returns all nodes that were ever a child of the parent node,
         at any point during the page's lifetime."""
-        if parent_node not in self.inserted_below_mapping:
+        if parent_node not in self.__inserted_below_map:
             return None
-        return self.inserted_below_mapping[parent_node]
+        return self.__inserted_below_map[parent_node]
 
     @lru_cache(maxsize=None)
     def node(self, node_id: PageGraphId) -> Node:
@@ -178,7 +201,7 @@ class PageGraph:
         node = node_for_type(node_type, self, node_id)
         if node.is_dom_node_type():
             dom_node = cast(DOMNode, node)
-            self.blink_id_mapping[dom_node.blink_id()] = dom_node
+            self.__blink_id_map[dom_node.blink_id()] = dom_node
         return node
 
     @lru_cache(maxsize=None)
@@ -186,7 +209,7 @@ class PageGraph:
         """Loading any edge object should come through this method, since
         this method is the one that knows what Edge or Edge subtype
         should be used."""
-        parent_id, child_id = self.edge_id_cache[edge_id]
+        parent_id, child_id = self.__edge_id_cache[edge_id]
         edge_key = (parent_id, child_id, edge_id)
         edge_data = self.graph.edges[edge_key]
         edge_type_str = edge_data[Edge.RawAttrs.TYPE.value]
@@ -199,9 +222,9 @@ class PageGraph:
             inserted_node = insert_edge.inserted_node()
             parent_node = insert_edge.inserted_below_node()
 
-            if parent_node not in self.inserted_below_mapping:
-                self.inserted_below_mapping[parent_node] = []
-            self.inserted_below_mapping[parent_node].append(inserted_node)
+            if parent_node not in self.__inserted_below_map:
+                self.__inserted_below_map[parent_node] = []
+            self.__inserted_below_map[parent_node].append(inserted_node)
         return edge
 
     def iframe_nodes(self) -> List[FrameOwnerNode]: