+++ pagegraph/graph/element.py
@@ -1,7 +1,8 @@
+from abc import abstractmethod
 import sys
 from typing import Any, Dict, TYPE_CHECKING
 
-from pagegraph.graph.types import PageGraphId
+from pagegraph.types import PageGraphId
 
 if TYPE_CHECKING:
     from pagegraph.graph import PageGraph
@@ -22,14 +23,24 @@ class PageGraphElement:
     def id(self) -> PageGraphId:
         return self._id
 
+    @abstractmethod
+    def validate(self) -> bool:
+        raise NotImplementedError()
+
+    @abstractmethod
     def data(self) -> Dict[str, Any]:
         raise NotImplementedError("Child class must implement 'data'")
 
+    @abstractmethod
     def timestamp(self) -> int:
         raise NotImplementedError("Child class must implement 'timestamp'")
 
+    @abstractmethod
     def describe(self) -> str:
         raise NotImplementedError("Child class must implement 'describe'")
+
+    def build_caches(self) -> None:
+        pass
 
     def throw(self, desc: str) -> None:
         sys.stderr.write(self.describe())